otp implementation

+++ src/app/login/component/login/login.component.ts
@@ -22,45 +22,41 @@ export class LoginComponent {
   isLogged = false;
   token!: string;
   otp!: string;
+  nextButtonIfOTPConfirmed!: any;
 
   constructor(private _formBuilder: FormBuilder, private _service: ApiService) {}
 
-  submitData() {
+  getOTP() {
     this.username = this.firstFormGroup.controls['firstCtrl'].value  as string;
     this.password = this.secondFormGroup.controls['secondCtrl'].value  as string;
+console.log("DADOS", this.username)
+    this._service.getOTP({username: this.username, password: this.password}).subscribe({
+      next: (OTP: any) => {
+        this.otp = OTP.OTP;
+        console.log("OTP", OTP);
+      },
+      error: (err) => console.log(err)
+    })
+  }
 
-    // @ts-ignore
-    this._service.login({username: this.username, password: this.password}).subscribe({
-      next: (data: Token) => {
-        this.isLogged = true;
+  confirmOTP(OTP: string) {
+    console.log("valor input", OTP)
+    this._service.confirmOTP({otp: OTP}).subscribe({
+      next: (data: any) => {
+        console.log("Confirmed token", data)
         this.token = data.token;
         localStorage.setItem('token', data.token)
-        console.log(data.token, this.isLogged)
+        this.nextButtonIfOTPConfirmed = document.getElementById('nextButtonOTP') as HTMLElement;
+        this.nextButtonIfOTPConfirmed.click();
+        this.isLogged = true;
+        this.token = data.token;
       },
-      error: (err: any) => {
-        this.isLogged = false;
-        console.log(err)
+      error: (err) => {
+        console.log(err),
+          this.nextButtonIfOTPConfirmed = document.getElementById('nextButtonOTP') as HTMLElement;
+        this.nextButtonIfOTPConfirmed.click();
+          this.isLogged = false;
       }
-    })
-  }
-
-  getOTP() {
-    this._service.getOTP().subscribe({
-      next: (OTP) => {
-        console.log("OTP", OTP)
-      },
-      error: (err) => console.log(err)
-    })
-  }
-
-  insertOTP(value: string) {
-    console.log("valor input", value)
-    this._service.sendOTP(value).subscribe({
-      next: (OTP: any) => {
-        console.log("OTP", OTP)
-        this.otp = OTP.OTP;
-      },
-      error: (err) => console.log(err)
     })
   }
 }
